refactor: refactor quotes into a separate file
this also fixes memory leaks
